requirements.txt
boto3==1.28.56   # Update this to the version you are using
python-dotenv==1.0.0  # Update this to the version you are using
